requirements.txt
python-telegram-bot==20.8
Pillow==10.2.0
wkhtmltopdf==0.2
pdfkit==1.0.0
imgkit==1.2.3
jinja2==3.1.3
requests==2.31.0
python-dotenv==1.0.0
cairosvg==2.7.1
